Rename InstallHandlerBtn to AllowWalletButton.vue.
